# perf: process nodes in lower bound order