This is for faster Branch-and-Bound solve times.

The benchmarks indicate that this works at least ok for the small instances. I think it is worth it since it sets the foundation for more sophisticated node selection or exploration strategies.

This commit:
BenchmarkBBOnRandomTinyInstances-4                    18          62534993 ns/op
BenchmarkBBOnRandomScale1TinyInstances-4              36          31242361 ns/op
BenchmarkBBOnRandomScale1000TinyInstances-4           36          31222192 ns/op
BenchmarkBBOnRandomSmallInstances-4                   12          93283916 ns/op

Previous commit:
BenchmarkBBOnRandomTinyInstances-4                    18          62894805 ns/op
BenchmarkBBOnRandomScale1TinyInstances-4              37          31532957 ns/op
BenchmarkBBOnRandomScale1000TinyInstances-4           36          32144553 ns/op
BenchmarkBBOnRandomSmallInstances-4                   12         108922024 ns/op